Add index menu/navigator and center recipe images
Index menu was added by making it into block  display and used padding
and margin to separate them in neat fashion. Added color to the fonts
and background to make them pop.

Images for each recipes were centered so the page looks more captivating
and easy to focus on with eyes in the center of the page.

[Ticket: X]